refactor: Update imports and remove unnecessary code in auth.ts
